refactor: extract colapse panels components
