+++ client/src/modules/Tasks/components/TaskSettings/TaskSettings.test.tsx
@@ -1,0 +1,27 @@
+import { render, screen } from '@testing-library/react';
+import { TaskSettings } from './TaskSettings';
+import { Form } from 'antd';
+import { TASK_SETTINGS_HEADERS } from 'modules/Tasks/constants';
+import { CriteriaDto } from 'api';
+
+const renderTaskSettings = (dataCriteria: CriteriaDto[] = [], setDataCriteria = jest.fn()) => {
+  render(
+    <Form>
+      <TaskSettings dataCriteria={dataCriteria} setDataCriteria={setDataCriteria} />
+    </Form>,
+  );
+};
+
+describe('TaskSettings', () => {
+  test.each`
+    header
+    ${TASK_SETTINGS_HEADERS.crossCheckCriteria}
+    ${TASK_SETTINGS_HEADERS.github}
+    ${TASK_SETTINGS_HEADERS.jsonAttributes}
+  `('should render task setting panel $header', ({ header }) => {
+    renderTaskSettings();
+
+    const panel = screen.getByText(header);
+    expect(panel).toBeInTheDocument();
+  });
+});
