+++ client/src/modules/Tasks/components/index.ts
@@ -1,2 +1,6 @@
 export * from './TableView/TableView';
 export * from './TaskModal/TaskModal';
+export * from './TaskSettings/TaskSettings';
+export * from './CrossCheckTaskCriteriaPanel/CrossCheckTaskCriteriaPanel';
+export * from './GitHubPanel/GitHubPanel';
+export * from './JsonAttributesPanel/JsonAttributesPanel';
